feat: NextAuth complete — DB user upsert on sign-in, webhook DB wiring, account settings
